server-gemini/requirements.txt
flask>=2.0
flask-cors>=3.0
google-genai>=0.4.0
gunicorn>=21.2.0